Transmitter code, first version posted!
Verbose transmission.

Basic transmitter functions added. Transmitter header file with all the
key information is available. Signal class is modified to include a
universal BitWait timing, to be later used by the Receiver class as
well.

Test file for the Transmitter is available, use an oscilloscope on the
pins with a pullup resistor to check signal outputs.

Transmitter_old is left in the folder for Nina and Kyu to use, otherwise
Test_Transmitter should also be functional.